Use Vec for PathRouter
Routes in path seemed to work with linear memory instead of needing the
extra hash and tracking of prev_route_id, a Vec should work just fine
while having the benefits of cache locality.

+++ axum/src/routing/mod.rs
@@ -58,7 +58,7 @@ macro_rules! panic_on_err {
 }
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub(crate) struct RouteId(u32);
+pub(crate) struct RouteId(usize);
 
 /// The router type for composing handlers and services.
 ///
